Refactor: Convert to vanilla Express
Convert to vanilla Express app to deploy into gcloud.

Resolves: N/a
See also: N/a
