+++ app.yaml
@@ -1,0 +1,8 @@
+runtime: nodejs10
+service: sherpa
+
+handlers:
+  - url: /.*
+    script: auto
+    secure: always
+    redirect_http_response_code: 301
