--- index.html
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-<object
-  type="text/html"
-  data="/.netlify/functions/server"
-  style="width:100%;height:100%;"
->
