--- netlify.toml
@@ -1,3 +1,0 @@
-[build]
-  command = "npm install && npm run build"
-  publish = "build"
