+++ package.json
@@ -2,7 +2,7 @@
   "name": "sherpabot",
   "version": "0.1.0",
   "description": "A Discord 'bot to help guide members of the Chingu Community",
-  "main": "src/server.js",
+  "main": "server.js",
   "directories": {
     "doc": "docs"
   },
@@ -21,16 +21,14 @@
   },
   "homepage": "https://github.com/chingu-x/sherpabot#readme",
   "scripts": {
-    "start": "nodemon server-local.js"
+    "build": "",
+    "start": "node server.js"
   },
   "dependencies": {
-    "discord.js": "^11.5.1"
-  },
-  "devDependencies": {
+    "discord.js": "^11.5.1",
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "nodemon": "^1.19.2",
-    "serverless-http": "^2.3.0"
+    "nodemon": "^1.19.2"
   }
 }
